fix(agent-gui): satisfy fork recovery lint
